+++ docs/CHANGELOG.md
@@ -25,3 +25,4 @@
     - **N.I.N.A 3.1 HF1**
 
 ## 2.1.1 2024-07-26 in work
+1. ALPACA optional focuser services implemented